fix: add edit modal in side panel in order to remove scroll behaivior
