+++ frontend/src/components/ActivityLogTable.tsx
@@ -19,14 +19,12 @@ import { InternalErrors } from '../common/errors/internal-errors.class';
 import MediaCell from './MediaCell';
 import {
   useActivityLogCounterQuery,
-  useActivityLogQuery,
   useActivityLogsQuery,
 } from '../services/activity-log/activity-log.service';
 import { IActivityLogListItem } from '../common/interfaces/activity-log.interface';
 import CellLayout from '../layouts/CellLayout';
 import StatusWithMarker from './StatusWithMarker';
 import { useNavigate } from 'react-router';
-import EditActivityLog from './EditActivityLog';
 import ActivityLogSidePanel from './ActivityLogSidePanel';
 import DataTableFilters from './DataTableFilters';
 import DateRangePicker from './DateRangePicker';
@@ -263,11 +261,9 @@ const ActivityLogTable = ({
   const [status, setStatus] = useState<SelectItem<ActivityLogStatus>>();
 
   // selected activity log id
-  const [selectedActivityLog, setSelectedActivityLog] = useState<string | null>(null);
+  const [selectedActivityLog, setSelectedActivityLog] = useState<string>();
   // side panel state
   const [isViewActivityLogSidePanelOpen, setIsViewActivityLogSidePanelOpen] =
-    useState<boolean>(false);
-  const [isEditctivityLogSidePanelOpen, setIsEditActivityLogSidePanelOpen] =
     useState<boolean>(false);
 
   // get all query
@@ -293,22 +289,13 @@ const ActivityLogTable = ({
   });
   const { data: counters } = useActivityLogCounterQuery(volunteerId);
 
-  // get one query
-  const { data: activityLog, error: activityLogError } = useActivityLogQuery(
-    selectedActivityLog as string,
-  );
-
   // query error handling
   useEffect(() => {
     if (activityLogsError)
       useErrorToast(
         InternalErrors.ACTIVITY_LOG_ERRORS.getError(activityLogsError.response?.data.code_error),
       );
-    if (activityLogError)
-      useErrorToast(
-        InternalErrors.ACTIVITY_LOG_ERRORS.getError(activityLogError.response?.data.code_error),
-      );
-  }, [activityLogsError, activityLogError]);
+  }, [activityLogsError]);
 
   // menu items
   const buildActivityLogActionColumn = (): TableColumn<IActivityLogListItem> => {
@@ -350,20 +337,9 @@ const ActivityLogTable = ({
     setIsViewActivityLogSidePanelOpen(true);
   };
 
-  const onEdit = () => {
-    setIsViewActivityLogSidePanelOpen(false);
-    setIsEditActivityLogSidePanelOpen(true);
-  };
-
-  const onCloseEditSidePanel = (shouldRefetch?: boolean) => {
-    setIsViewActivityLogSidePanelOpen(true);
-    setIsEditActivityLogSidePanelOpen(false);
-    if (shouldRefetch) refetch();
-  };
-
   const onCloseSidePanel = (shouldRefetch?: boolean) => {
     setIsViewActivityLogSidePanelOpen(false);
-    setSelectedActivityLog(null);
+    setSelectedActivityLog(undefined);
     if (shouldRefetch) refetch();
   };
 
@@ -553,14 +529,8 @@ const ActivityLogTable = ({
       </Card>
       <ActivityLogSidePanel
         onClose={onCloseSidePanel}
-        onEdit={onEdit}
         isOpen={isViewActivityLogSidePanelOpen}
-        activityLog={activityLog}
-      />
-      <EditActivityLog
-        onClose={onCloseEditSidePanel}
-        isOpen={isEditctivityLogSidePanelOpen}
-        activityLog={activityLog}
+        activityLogId={selectedActivityLog}
       />
     </>
   );
